create commponents/layout folder and move files
